Move things around, remove dead code
