+++ io_scene_hubs/__init__.py
@@ -4,23 +4,23 @@ from . import settings
 from . import components
 from . import operators
 from . import panels
-
-from io_scene_gltf2.blender.exp import gltf2_blender_export
-from io_scene_gltf2.io.exp.gltf2_io_user_extensions import export_user_extensions
+from .gather_properties import gather_properties
 
 bl_info = {
     "name" : "io_scene_hubs",
     "author" : "Robert Long",
     "description" : "",
-    "blender" : (2, 80, 0),
-    "version" : (0, 0, 1),
+    "blender" : (2, 82, 0),
+    "version" : (0, 0, 2),
     "location" : "",
     "warning" : "",
     "category" : "Generic"
 }
 
+# Monkey patch to add gather_gltf_hook, can be removed once handled upstream https://github.com/KhronosGroup/glTF-Blender-IO/issues/1009
+from io_scene_gltf2.blender.exp import gltf2_blender_export
+from io_scene_gltf2.io.exp.gltf2_io_user_extensions import export_user_extensions
 orig_gather_gltf = gltf2_blender_export.__gather_gltf
-
 def patched_gather_gltf(exporter, export_settings):
     orig_gather_gltf(exporter, export_settings)
     export_user_extensions('gather_gltf_hook', export_settings, exporter._GlTF2Exporter__gltf)
@@ -34,8 +34,6 @@ def register():
     operators.register()
     panels.register()
 
-    bpy.utils.register_class(HubsComponentsExtensionProperties)
-    bpy.types.Scene.HubsComponentsExtensionProperties = bpy.props.PointerProperty(type=HubsComponentsExtensionProperties)
 
 def unregister():
     gltf2_blender_export.__gather_gltf = orig_gather_gltf
@@ -45,85 +43,28 @@ def unregister():
     operators.unregister()
     panels.unregister()
 
-    unregister_panel()
-    bpy.utils.unregister_class(HubsComponentsExtensionProperties)
-    del bpy.types.Scene.HubsComponentsExtensionProperties
+    unregister_export_panel()
 
 if __name__ == "__main__":
     register()
 
-class HubsComponentsExtensionProperties(bpy.types.PropertyGroup):
-    enabled: bpy.props.BoolProperty(
-        name="Export Hubs Components",
-        description='Include this extension in the exported glTF file.',
-        default=True
-        )
-    float_property: bpy.props.FloatProperty(
-        name='Sample FloatProperty',
-        description='This is an example of a FloatProperty used by a UserExtension.',
-        default=1.0
-        )
-
-
 # called by gltf-blender-io after it has loaded
 def register_panel():
-    # Register the panel on demand, we need to be sure to only register it once
-    # This is necessary because the panel is a child of the extensions panel,
-    # which may not be registered when we try to register this extension
     try:
-        bpy.utils.register_class(GLTF_PT_UserExtensionPanel)
+        bpy.utils.register_class(panels.HubsGLTFExportPanel)
     except Exception:
         pass
+    return unregister_export_panel
 
-    # If the glTF exporter is disabled, we need to unregister the extension panel
-    # Just return a function to the exporter so it can unregister the panel
-    return unregister_panel
-
-
-def unregister_panel():
+def unregister_export_panel():
     # Since panel is registered on demand, it is possible it is not registered
     try:
-        bpy.utils.unregister_class(GLTF_PT_UserExtensionPanel)
+        bpy.utils.unregister_class(panels.HubsGLTFExportPanel)
     except Exception:
         pass
 
-class GLTF_PT_UserExtensionPanel(bpy.types.Panel):
-
-    bl_space_type = 'FILE_BROWSER'
-    bl_region_type = 'TOOL_PROPS'
-    bl_label = "Hubs Components"
-    bl_parent_id = "GLTF_PT_export_user_extensions"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    @classmethod
-    def poll(cls, context):
-        sfile = context.space_data
-        operator = sfile.active_operator
-        return operator.bl_idname == "EXPORT_SCENE_OT_gltf"
-
-    def draw_header(self, context):
-        props = bpy.context.scene.HubsComponentsExtensionProperties
-        self.layout.prop(props, 'enabled', text="")
-
-    def draw(self, context):
-        layout = self.layout
-        layout.use_property_split = True
-        layout.use_property_decorate = False  # No animation.
-
-        props = bpy.context.scene.HubsComponentsExtensionProperties
-        layout.active = props.enabled
-
-        box = layout.box()
-        box.label(text="No options yet")
-
-
-import logging
-
-log = logging.getLogger(__name__)
-from . import exporter
-
+# This class name is specifically looked for by gltf-blender-io and it's hooks are automatically invoked on export
 class glTF2ExportUserExtension:
-
     def __init__(self):
         # We need to wait until we create the gltf2UserExtension to import the gltf2 modules
         # Otherwise, it may fail because the gltf2 may not be loaded yet
@@ -165,7 +106,7 @@ class glTF2ExportUserExtension:
                 component_class = registered_hubs_components[component_name]
                 component_class_name = component_class.__name__
                 component = getattr(blender_object, component_class_name)
-                component_data[component_name] = exporter.gather_properties(export_settings, blender_object, component, component_definition, hubs_config)
+                component_data[component_name] = gather_properties(export_settings, blender_object, component, component_definition, hubs_config)
 
             if gltf2_object.extensions is None:
                 gltf2_object.extensions = {}
