+++ io_scene_hubs/components.py
@@ -18,6 +18,18 @@ class HubsComponentName(PropertyGroup):
 class HubsComponentList(PropertyGroup):
     items: CollectionProperty(type=HubsComponentName)
 
+class HubsComponentsExtensionProperties(bpy.types.PropertyGroup):
+    enabled: bpy.props.BoolProperty(
+        name="Export Hubs Components",
+        description='Include this extension in the exported glTF file.',
+        default=True
+        )
+    float_property: bpy.props.FloatProperty(
+        name='Sample FloatProperty',
+        description='This is an example of a FloatProperty used by a UserExtension.',
+        default=1.0
+        )
+
 def is_scene_component(component_definition):
     return 'scene' in component_definition and component_definition['scene']
 
@@ -242,9 +254,11 @@ def register():
     bpy.utils.register_class(Material2DArrayValueProperty)
     bpy.utils.register_class(HubsComponentName)
     bpy.utils.register_class(HubsComponentList)
+    bpy.utils.register_class(HubsComponentsExtensionProperties)
     bpy.types.Object.hubs_component_list = PointerProperty(type=HubsComponentList)
     bpy.types.Scene.hubs_component_list = PointerProperty(type=HubsComponentList)
     bpy.types.Material.hubs_component_list = PointerProperty(type=HubsComponentList)
+    bpy.types.Scene.HubsComponentsExtensionProperties = PointerProperty(type=HubsComponentsExtensionProperties)
 
 def unregister():
     bpy.utils.unregister_class(StringArrayValueProperty)
@@ -252,6 +266,8 @@ def unregister():
     bpy.utils.unregister_class(Material2DArrayValueProperty)
     bpy.utils.unregister_class(HubsComponentName)
     bpy.utils.unregister_class(HubsComponentList)
+    bpy.utils.unregister_class(HubsComponentsExtensionProperties)
     del bpy.types.Object.hubs_component_list
     del bpy.types.Scene.hubs_component_list
     del bpy.types.Material.hubs_component_list
+    del bpy.types.Scene.HubsComponentsExtensionProperties
